refactor(map): consolidate map hooks into MapSection component
