--- apps/web/src/hooks/map/index.ts
@@ -1,3 +1,0 @@
-export * from './useMapBounds'
-export * from './useMapMarkerSelection'
-export * from './usePhotoMarkers'
